requirements.txt
streamlit==1.28.1
google-generativeai==0.4.0
reportlab==4.0.4
python-dotenv==1.0.0
